requirements.txt
nose==1.3.7
sphinx==1.8.1
python-dotenv==0.9.1
requests==2.19.1
quantconnect==0.1
beautifulsoup4==4.6.3
lxml==4.2.5
pandas==0.23.4
